requirements.txt
Pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers